Do not use index id in processes Fix tools CLI Remove provisioning plugin workaround
