--- etc/src/index.js
@@ -1,4 +1,0 @@
-'use strict';
-
-// There is nothing here
-
