+++ lib/cf/dashboard/package.json
@@ -25,7 +25,7 @@
   "scripts": {
     "postinstall": "cd src/webapp; yarn install; cd ../..;",
     "start": "node src/index.js &",
-    "stop": "pkill -f 'node abacus-service-dashboard-[0-9]+-[0-9]+ express' || true",
+    "stop": "pkill -f 'node abacus-service-dashboard-[0-9]+ express' || true",
     "lint": "abacus-eslint",
     "pub": "publish",
     "cfpush": "cfpush",
