+++ lib/plugins/provisioning/src/index.js
@@ -9,16 +9,12 @@
 // This minimal Abacus provisioning plugin example is provided only for demo
 // and test purposes. An integrator of Abacus is expected to replace it with
 // a real production implementation.
-const fs = require('fs');
-const path = require('path');
-
 const { extend, isEmpty, pick } = require('underscore');
 
 const oauth = require('abacus-oauth');
 const mappings = require('abacus-plan-mappings');
 const router = require('abacus-router');
 const webapp = require('abacus-webapp');
-const yieldable = require('abacus-yieldable');
 
 const plansDb = require('./lib/plan-db.js');
 
@@ -43,42 +39,6 @@ const rtype = function*(rid) {
 
 const writeScope = { system: ['abacus.usage.write'] };
 const readScope = { system: ['abacus.usage.read'] };
-
-const upsertPlan = function*(type, plan) {
-  if (yield plansDb.read({ cache: false })[type](plan.plan_id))
-    yield plansDb.update[type](plan.plan_id, plan);
-  else
-    yield plansDb.create[type](plan);
-};
-
-const storeAllPlansOfType = (type, cb) => {
-  const planDir = path.join(__dirname, '/plans/' + type);
-  const planFiles = fs.readdirSync(planDir);
-  yieldable.functioncb(function*() {
-    for(let file in planFiles) {
-      const defaultPlan = require(path.join(planDir, planFiles[file]));
-      yield upsertPlan(type, defaultPlan);
-    }
-  })((error) => {
-    if(error)
-      throw new Error('Failed to store default plan: ' + error);
-
-    debug('Default %s plan mappings created', type);
-    cb();
-  });
-};
-
-const storeAllDefaultPlans = (cb = () => {}) => {
-  let callCount = 0;
-  const countCb = () => {
-    if(++callCount === 3)
-      cb();
-  };
-  const types = ['metering', 'pricing', 'rating'];
-
-  for(let type of types)
-    storeAllPlansOfType(type, countCb);
-};
 
 const throwUnauthorizedError = (scopes, planId) => {
   edebug('Provided scopes %j are insufficient to access plan: %s', scopes,
@@ -465,7 +425,5 @@ const runCLI = () => provisioning().listen();
 
 // Export our public functions
 module.exports = provisioning;
-module.exports.storeAllDefaultPlans = storeAllDefaultPlans;
 
-if (!process.env.NON_INTERACTIVE)
-  runCLI();
+runCLI();
