+++ lib/plugins/provisioning/store-default-plans
@@ -1,10 +1,8 @@
 #!/usr/bin/env node
 'use strict';
 
-process.env.NON_INTERACTIVE = true;
-
 const debug = require('abacus-debug')('abacus-provisioning-plugin');
-const provisioning = require('./src/index');
+const provisioning = require('./src/lib/store-plans.js');
 
 provisioning.storeAllDefaultPlans(() => {
   debug('Default plans have been imported');
