+++ lib/sampler/receiver/src/index.js
@@ -129,13 +129,13 @@ const startApp = async () => {
   await listen(app);
 };
 
-module.exports = {
-  runCLI: () => {
-    startApp().then(() => {
-      debug('Receiver web application is successfully started.');
-    }).catch((err) => {
-      edebug('Error while starting receiver web application. Error: %o', err);
-      process.exit(1);
-    });
-  }
+const runCLI = () => {
+  startApp().then(() => {
+    debug('Receiver web application is successfully started.');
+  }).catch((err) => {
+    edebug('Error while starting receiver web application. Error: %o', err);
+    process.exit(1);
+  });
 };
+
+runCLI();
