+++ lib/sampler/worker/src/index.js
@@ -111,13 +111,13 @@ const startApp = async () => {
   await listen(app);
 };
 
-module.exports = {
-  runCLI: () => {
-    startApp().then(() => {
-      debug('Worker web application is successfully started.');
-    }).catch((err) => {
-      edebug('Error while starting worker web application. Error: %o', err);
-      process.exit(1);
-    });
-  }
+const runCLI = () => {
+  startApp().then(() => {
+    debug('Worker web application is successfully started.');
+  }).catch((err) => {
+    edebug('Error while starting worker web application. Error: %o', err);
+    process.exit(1);
+  });
 };
+
+runCLI();
