+++ lib/utils/express/src/index.js
@@ -98,7 +98,7 @@ const truncate = (s, length) => {
 // common text-based mime types
 const isText = (mime) => {
   return new RegExp(
-    'css|form-data|html|javascript|json|perl|postscript|rdf|' + 'rtf|sgml|svg|text|urlencoded|xml'
+    'css|form-data|html|javascript|json|perl|postscript|rdf|rtf|sgml|svg|text|urlencoded|xml'
   ).test(mime);
 };
 
@@ -470,7 +470,7 @@ const listen = function(opt, cb) {
   const conf = lconf(opt);
 
   // Set the process title
-  const vcapAppTitle = [vcapenv.appname(), vcapenv.appindex(), vcapenv.iid()].join('-');
+  const vcapAppTitle = [vcapenv.appname(), vcapenv.appindex()].join('-');
   process.title = `node ${process.env.TITLE || vcapAppTitle} express`;
 
   // Register some of our middleware after the routes
