+++ tools/cfpush/src/index.js
@@ -82,5 +82,3 @@ const runCLI = () => {
 };
 
 module.exports.runCLI = runCLI;
-
-runCLI();
