+++ tools/mddoc/src/index.js
@@ -90,4 +90,4 @@ const runCLI = (stdin, stdout) => {
   });
 };
 
-runCLI();
+module.exports.runCLI = runCLI;
